+++ pkg/frontend/v1/frontend_test.go
@@ -40,6 +40,7 @@ import (
 	"github.com/grafana/mimir/pkg/querier/stats"
 	querier_worker "github.com/grafana/mimir/pkg/querier/worker"
 	"github.com/grafana/mimir/pkg/scheduler/queue"
+	util_test "github.com/grafana/mimir/pkg/util/test"
 )
 
 const (
@@ -222,10 +223,7 @@ func TestFrontendMetricsCleanup(t *testing.T) {
 
 		fr.cleanupInactiveUserMetrics("1")
 
-		require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
-				# HELP cortex_query_frontend_queue_length Number of queries in the queue.
-				# TYPE cortex_query_frontend_queue_length gauge
-			`), "cortex_query_frontend_queue_length"))
+		util_test.AssertGatherAndCompare(t, reg, "", "cortex_query_frontend_queue_length")
 	}
 
 	testFrontend(t, defaultFrontendConfig(), handler, test, nil, reg)