+++ pkg/storage/ingest/partition_offset_client_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/twmb/franz-go/pkg/kmsg"
 	"go.uber.org/atomic"
 
+	"github.com/grafana/mimir/pkg/util/test"
 	"github.com/grafana/mimir/pkg/util/testkafka"
 )
 
@@ -381,7 +382,7 @@ func TestPartitionOffsetClient_FetchPartitionsLastProducedOffsets(t *testing.T) 
 			# HELP cortex_ingest_storage_reader_last_produced_offset_failures_total Total number of failed requests to get the last produced offset.
 			# TYPE cortex_ingest_storage_reader_last_produced_offset_failures_total counter
 			cortex_ingest_storage_reader_last_produced_offset_failures_total{partition="mixed"} 0
-		
+
 			# HELP cortex_ingest_storage_reader_last_produced_offset_requests_total Total number of requests issued to get the last produced offset.
 			# TYPE cortex_ingest_storage_reader_last_produced_offset_requests_total counter
 			cortex_ingest_storage_reader_last_produced_offset_requests_total{partition="mixed"} 4
@@ -409,9 +410,9 @@ func TestPartitionOffsetClient_FetchPartitionsLastProducedOffsets(t *testing.T) 
 		require.NoError(t, err)
 		assert.Empty(t, offsets)
 
-		assert.NoError(t, promtest.GatherAndCompare(reg, strings.NewReader(""),
+		test.AssertGatherAndCompare(t, reg, "",
 			"cortex_ingest_storage_reader_last_produced_offset_requests_total",
-			"cortex_ingest_storage_reader_last_produced_offset_failures_total"))
+			"cortex_ingest_storage_reader_last_produced_offset_failures_total")
 	})
 
 	t.Run("should honor context deadline and not fail other in-flight requests issued while the canceled one was still running", func(t *testing.T) {